+++ Makefile
@@ -26,10 +26,11 @@ EXEC_FILES = exec.c absolute_path.c fd_manager.c \
 				utils.c redirect_handler.c
 EXEC = $(addprefix $(EXEC_DIR), $(EXEC_FILES))
 
-SRC_FILES = main.c free_all.c
-SRC = $(addprefix $(SRC_DIR), $(SRC_FILES))  $(PARSING) $(EXEC) $(BUILTINS) $(ENV)
+MAIN_DIR = $(SRC_DIR)main/
+MAIN_FILES = main.c free_all.c simple_parser.c prompt_utils.c
+MAIN = $(addprefix $(MAIN_DIR), $(MAIN_FILES))
 
-
+SRC =  $(MAIN) $(PARSING) $(EXEC) $(BUILTINS) $(ENV)
 
 
 OBJ_DIR = ./obj/
