+++ includes/types.h
@@ -6,7 +6,7 @@
 /*   By: malde-ch <malo@chato.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 05:44:09 by malde-ch          #+#    #+#             */
-/*   Updated: 2025/03/04 18:53:21 by malde-ch         ###   ########.fr       */
+/*   Updated: 2025/03/06 18:07:42 by malde-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,23 +24,19 @@ typedef enum e_operator
 {
 	OP_NONE, // 0
 	OP_PIPE, // |
-
-
 	OP_REDIRECT_OUT, // >
 	OP_REDIRECT_OUT_APPEND, // >>
 	OP_REDIRECT_IN, // <
 	OP_HERE_DOC, // <<
-
-	
 	OPERATOR,
 }	t_operator;
 
 typedef struct s_cmd
 {
-	char		**cmd;
-	int			fd_in;
-	int			fd_out;
-	t_operator	operator;
+	char			**cmd;
+	int				fd_in;
+	int				fd_out;
+	t_operator		operator;
 	struct s_cmd	*next;
 }	t_cmd;
 
@@ -51,7 +47,6 @@ typedef struct s_mini
 	t_env_var	*env;
 	t_cmd		*cmd;
 }	t_mini;
-
 
 /* 
 	Cela est un alias de la fonction builtin
@@ -68,13 +63,12 @@ typedef struct s_builtin
 	t_builtin_func	func;
 }	t_builtin;
 
-typedef int (*t_handler_func)(t_mini *mini, t_cmd *cmd, t_cmd *op);
+typedef int	(*t_handler_func)(t_mini *mini, t_cmd *cmd, t_cmd *op);
 
 typedef struct s_handler
 {
-	t_operator	operator;
+	t_operator		operator;
 	t_handler_func	func;
 }	t_handler;
-
 
 #endif
